LPS-82004 portal-search-test-util: Move duplicate logic to base class
https://issues.liferay.com/browse/LPS-82004
